jobs: retire osmocom-gerrit-debian* labels
Let the gerrit-binpkgs and gerrit-pipeline-result jobs use the
"osmocom-gerrit" label. Having the -debian10/11 label previously was
useful to make sure this doesn't run on an ancient python version that
fails on f-strings. But since the debian9 builders were retired, this is
not needed anymore. These are the only jobs that used these, so now the
labels in the README can be simplified (following patch).

Related: OS#5793
Change-Id: Ieee95d25e036dcd4869d89b10c281179024672e7