Revert "jobs: tweak build-discarder values"
This reverts commit 719ff9760808acb1fe944f4c496c356c7dc07c91.

The disk consumption of jenkins has grown by almost 500GB
since we merged this patch.  Clearly this is not expected, and we'd
have ran out of disk space in a few weeks.

I personally think the current allocation of 1.5TB of disk space to
jenkins should be more than sufficient; we just need to manage it
better.

Closes: OS#5980
Change-Id: I6b744a8b84a3e1255a8d51f73d1721ccfd028ac1