obs: update_obs_project: add ALL_OSMOCOM_PACKAGES
Add this alias to use all Osmocom packages, but no the misc packages
that would also be selected if not specifying the packages argument at
all (limesuite, neocon, open5gs).

This will be used for building packages from the rhizomatica branches,
so we don't check the git repositories of these misc projects for the
git branches.

Related: OS#5981
Change-Id: I4265f43408bc21ac8765449d6766381c75cafe86