data/build_rpm: force refresh of package indexes
The yum-builddep command already has a mechanism to update the package
index data if the last update (when the docker image was built) was some
time ago. However when the packages we use happened to be updated in
that timeframe, the indexes are not updated and attempting to download
them leads to 404 errors. Force refresh all package indexes to avoid
this.

Fixes: OS#6038
Change-Id: I6e2592466b6190782cdd93e00b3d5020fd665372