Fix failing Osmocom-release-manuals job
Add a workaround that upgrades the osmocom-nightly package before
installing build dependencies for the given package.

This is not very elegant, it would make more sense if the docker image
we use here did not have the nightly Osmocom repository configured in
the first place (as this job is about creating manuals for tagged
releases). But the image is used by jenkins build verification too and I
don't think it's a good use of time to change this right now.

Fix for:
  + apt-get -y build-dep /build
  Note, using directory '/build' to get the build dependencies
  Reading package lists...
  Building dependency tree...
  Reading state information...
  Some packages could not be installed. This may mean that you have
  requested an impossible situation or if you are using the unstable
  distribution that some required packages have not yet been created
  or been moved out of Incoming.
  The following information may help to resolve the situation:

  The following packages have unmet dependencies:
   osmo-gsm-manuals-dev : Depends: osmocom-nightly (= 202305300026) but 202305290026 is to be installed

Change-Id: If28a34d3e5b07216c5310b19623fcc42692f65c3