ansible: osmocom_build_deps: remove
The builds for almost all jenkins jobs are done in docker now, we
definitely don't need to install all of these dependencies on all
jenkins nodes anymore.

Remove the file for now, so it doesn't look like they still get used and
so we don't need to keep it up-to-date (e.g. inkscape was just replaced
with rsvg-convert).

When we deploy a new jenkins node and notice that we still do need a few
of these dependencies, we can bring them back later.

Change-Id: I7c87bd516d1bcdbcc21f65909dabbd3ccc38eede