ansible: increase docker_max_image_space
The purpose of the limit is to remove images that are not being used in
a long time. With the current limit of 40 GB it leads to images being
removed just a few hours after they were last used, resulting in lots
of needless rebuilds. Also unexpected errors with images that we just
assume to be there (debian-bookworm-build).

Increases the limit for the nodes that run most of the jobs to 80 GB.

I have verified that the nodes have enough free space.

Related: OS#6183
Change-Id: I8cfb21a86510e43b8347c7f17159575b695ce9ba