jobs: gerrit: build osmo-pcap + dep for centos7
We are about to remove the rpm spec files for almost all Osmocom git
repositories. The only two git repositories where we still need them,
because we have a customer using them on centos 7, are libosmocore and
osmo-pcap (SYS#6760). Add CI to ensure that these still build fine, and
that we don't remove these by mistake.

Related: OS#6446
Change-Id: I4d4a67c053ed29811c504158d1ef49430880ad4f