jenkins-jobs-osmo: new file
Do not recommend to the user anymore to put the Osmocom jenkins server
in their global jenkins-job-builder config, as this may lead to jobs
from other repositories being deployed to the Osmocom jenkins by
accident.

Instead of that, let the user put the config in this path:
	~/.config/jenkins_jobs/jenkins_jobs_osmo-ci.ini

Add a wrapper script that users this path, and makes sure that no global
config is present.

Related: SYS#6957
Change-Id: I1d7c50b2fef56e343cfe377b51f5f4304219cd22