gerrit-verifications-dahdi: set concurrent: false
Each of the gerrit-dahdi-linux-* jobs has a 3-4 GiB workspace directory
with a full linux tree. We keep them on disk so we don't need to clone
the linux tree with each build.

This makes sense, but set concurrent to false to not create more of
these than necessary. These directories stood out when investigating
why build4 was running out of space.

Change-Id: I4fd4e85f0930b15b6c67e70fb2c140392c910bf5