jobs/osmocom-build-tags…: fix workspace clean
As this job tries to archive the build logs after running, don't remove
the entire workspace. Instead remove the paths that take up the most
space, which are the one with the git clones and build artifacts (_deps)
as well as the one with the temporary installation dirs (_temp/_inst*).

Fixes: b20e8e8b ("jobs/various: clean workspace on success")
Change-Id: Ied86ba80b38a9b10a7a96631e6dedee9f3762c3b