OBS: osmocom:master: ignore missing commit_*.txt
Do not abort if the commit_*.txt file has been deleted from a package in
osmocom:master. Instead, assume the package is outdated, create a new
source package (with commit_*.txt) and replace the existing source
package.

This is useful to force rebuilding a package, e.g. if the previous
source package was built when the git tag was not pushed yet, and
therefore the version in the package and for libraries, in the
pkg-config data, don't have the proper version yet.

Change-Id: Ia6129f533394b94d5d8525eb0513a38a6e1eb7b6