OBS: fix up-to-date check for strongswan-epdg
Move the code to transform the version from the git tag to the version
that we can use in debian/changelog to a separate function. Use it not
only when generating the version for debian/changelog, but also when
checking if the current git tag is up-to-date.

Without this patch, it always considers strongswan-epdg outdated:
  strongswan-epdg: building source package (outdated: osmo-epdg-0.1.2 <=> 0.1.2.osmo.epdg in OBS)

Change-Id: I515695d52f5d37b256ef163875417745897e5900