jobs/{master,gerrit}: rpi4-raspbian11 -> raspbian
Rename the jenkins label rpi4-raspbian11 to raspbian so the name is not
misleading with how we currently use it. One raspberry pi has been
reinstalled with raspbian12 (OS#6549) and we also use it for LX2 with a
raspbian container (OS#5873, therefore remove rpi4 from the name).

The OS running directly on the rpi / in the LXC is not that important
anyway since we run the actual jobs in docker with the latest debian
version (currently 12).

Change-Id: I02dcf3098ffd3fa247c4f43c290d6e3fd809ad45