jobs/osmo-gsm-tester-build: fix Missing parameter
Use OSMO_GSM_TESTER_BRANCH as shell variable ('$...'), not as jenkins
job builder parameter (${...}). This is how we do it in other jobs as
well, e.g. jobs/osmocom-obs.yml.

Apparently the workaround in the file happened to work earlier, but it
does not work anymore with Jenkins Job Builder version 6.3.0. Remove it.

Fix for:
  osmo-gsm-tester-builder.yml:119:15: While formatting string '${OSMO_GSM_TESTER_BRANCH}': Missing parameter: 'OSMO_GSM_TESTER_BRANCH'
                - ${OSMO_GSM_TESTER_BRANCH}
                  ^

Change-Id: I3b6b16964a4d8a51fd3ecc0d670805182a8c6586