jobs/build-kernels-testenv: new jobs
Prepare to replace the docker-playground based jobs for running
osmo-ggsn in QEMU with a kernel built from source with testenv.

Split the part that builds the kernels into extra jobs that export the
built kernels as artifacts. This makes it easier to use kernels outside
of jenkins too and it allows retriggering only a testsuite run without
also building the kernel again.

The fragment.config is imported from docker-playground current master
978adc, path: ttcn3-ggsn-test/osmo-ggsn-kernel/fragment.config

Change-Id: Id64a1a778fa38eec20498c36b390332f75d7d3f5