jobs/build-kernels-testenv: add build-kernel-none
Create a dummy output/linux file, that the jobs generated by
ttcn3-testsuites-testenv can copy (via copyartifact) when setting the
kernel to "none".

Without this, the ttcn3-testsuites-testenv.yml would need to become more
complex, as it is not possible in jenkins job builder templates to only
do the copyartifact block conditionally if kernel != "none".

Another solution would be using shell logic like the following:

  if [ "$KERNEL" != "none" ]; then
    wget -q -O .linux "https://jenkins.osmocom.org/..."
  fi

But this has the disadvantage that jenkins is not aware of the file and
therefore we cannot track easily which linux kernel build was used in
the job. With the copyartifact method, we can click on "See
Fingerprints" and find which job built the kernel. Example:
https://jenkins.osmocom.org/jenkins/job/ttcn3-ggsn-test-kernel-net-next/147/fingerprints/

Change-Id: I0b2367558a2b73c06267bc65f17cfcc224f35002