jobs/ttcn3-testsuites-testenv: tweak description
Don't link to osmo-ttcn3-hacks.git directly anymore, the wiki page that
is also linked already refers to it. This shortens the boilerplate of
the description with every one of these jenkins jobs.

Instead of only displaying the arguments, display the full command so
users can easily run the testsuite the same way it is running on
jenkins. Omit the arguments --cache, --ccache, --log-dir, which only
need to be adjusted for jenkins, for users the defaults make sense.

Let the kernel line link to the related build-kernel job.

Related: OS#6494
Change-Id: Ifad86aa3982c325394d47b5c8cf98c43b2d3d1eb