jobs/ttcn3-testsuites: adjust timer for bts
BTS testsuites tend to fail with timeouts on heavy loads. Group them in
a timeslot where jenkins nodes are otherwise not busy according to
stats in grafana to hopefully make the test results more stable. Once
we can replace faketrx with a rewrite (OS#6672), this should not be
necessary anymore.

Change-Id: If25009a5b9215a0f51381529e26bfb3ba2303ae2