jobs: manually trigger "git gc" for linux.git
We already use shallow clones for linux.git, and git runs a garbage
collector automatically to ensure no longer needed artifacts are
removed. But this did not work anymore on build4 for
build-kernel-net-next, and it increased up to 8.4 GiB:

    8,4 GiB /build-kernel-net-next

After running "git gc" again:

    3,3 GiB /build-kernel-net-next

The reason was that "git gc" did not exit successfully once and wrote a
log file, then refused to run again. Run "git gc --auto" explicitly in
foreground to let abort in such a case, so it does not go unnoticed in
the future.

Change-Id: Ic81ae5a1edb0dbe119e06cc3b49ea69ef81b2a5f