lint/lint_diff: allow skipping with LINT=0
When this script is set as pre-commit hook, it can be useful to skip the
linter with this env var to make a commit that intentionally does not
pass the linter (e.g. to test if it will be caught in gerrit + jenkins).

There is also "git commit --no-verify", but with that the hook that
applies the Change-Id does not run.

Change-Id: Ia4563c13d9916a879a1e8afeacad8c541662adaa