obs: gerrit_binpkgs: set --allow-unknown-package
The gerrit_binpkgs.sh script can be used to quickly test building of
debian packages in a container even for packages that we don't have
in the OBS osmocom repositories. For example I'm currently using it to
test-build kamailio packages. Use the --allow-unknown-package argument
so build_srcpkg.py doesn't complain if this gets called with a package
that isn't a known Osmocom package.

Change-Id: I79e89c0d7a73f8b2262fd10a0f7a118aa8351635