ansible: build-hosts: add testenv-coredump-helper
The Osmocom jenkins nodes run inside LXCs. When we get a coredump it
appears on the host. Add a helper script to the hosts so the jenkins
jobs can fetch the coredumps in case an Osmocom program crashes while
running a ttcn3 testsuite.

The helper script has the following safety features to ensure jenkins
can't just fetch any coredump:

* Only fetch coredumps within the last 3 seconds and only if the
  executable matches osmo-* or open5gs-*
* Only listen on the lxc IP

Related: OS#6769
Change-Id: I7e66c98106b7028a393e3b873e96ae2dcb412c48