jobs: run ttcn3-bts-test-* between 01:00 - 06:00
The automatic respawns of osmo-bts-trx have been intentionally removed
in the related docker-playground patch, so we can better notice when
osmo-bts-trx crashes instead of just restarting it and moving on with
the testsuite.

The result is that now it is even more important when ttcn3-bts-test
jobs are running, because when the builders are under high load then
faketrx cannot send the time quickly enough and it also causes
osmo-bts-trx to shutdown.

In a discussion with Vadim we've decided to schedule the ttcn3-bts-tests
so that they don't only run at a time where no other jobs are running.
But also that they don't run at the same time on the same builder
anymore to give them the least unrelated load possible. Other solutions
are also being investigated, but this should make the job in its current
form more stable already.

Move misc ttcn3 jobs to run from 06:00 to 08:00 and run ttcn3-bts-tests
between 01:00 and 06:00. Pin them to specific runners to make sure we
don't have both running on the same node.

Related: docker-playground I770a0530f90c5e03d2073bfc1c4cfd0e840379c9
Change-Id: I0162f7299c8e37f893ffa10ddc4c8edece29ed7f