OBS: add the new osmocom-keyring package
The public key for OBS packages needs to be extended every two years.
When the previous key expires, it is not possible to update packages
anymore until users replace the old key with the new, extended key.

To make this less inconvenient for users, introduce a new
osmocom-keyring package that holds the current key (valid for 2 years),
which will get updated every year with an extended key (valid for
another 2 years, I'll add a jenkins job to remind us). This means as
long as users upgrade their packages once a year, they will
automatically receive the new key.

Related: OS#6567
Change-Id: I370b4fa0c6f7e8a925b1ddd7c1c47980a8f4fea2