OBS: let meta package depend on osmocom-keyring
Let the meta package (osmocom-latest, osmocom-nightly etc.) depend on
osmocom-keyring. This means that after the user has enabled any of the
official Osmocom OBS repositories and installs a package from there, the
osmocom-keyring package will get installed as well, which results in
having an up-to-date /usr/share/osmocom-keyring/osmocom.asc.

Change-Id: Id4d213f355338b39a9ec5ccffc16179f6e472102