jobs: make references to master branch consistent
Use "master" instead of:
- "refs/remotes/origin/master"
- "origin/master"
- "*/master"

This improves consistency and makes clear what should be used when
adding new jobs. It also fixes the problem with */master that it may
actually not only use the master branch, but also e.g. osmith/master
which can lead to unintentionally building and publishing artifacts from
the wrong branch. If we want to run a job on a different branch
intentionally, such a job should have an extra argument for branch that
can be changed when starting the job.

Change-Id: I7389585165bf3a790c7a338d10f5f7fe2311ee69