jobs: run ttcn3-remsim-test-asan with bookworm
The default distribution to run ttcn3 testsuites and SUTs with has
recently been changed from debian 12 (bookworm) to 13 (trixie). Since
then we have seen 27 additional failures with osmocom:nightly:asan
(while osmocom:nightly and osmocom:latest work fine). As the reason for
this is currently unclear, revert to using debian 12 for this job for
now, so this doesn't mask other errors that the job would find
otherwise.

Related: OS#6845
Change-Id: I702770e36b47df0a47bc9a869127d6251290a7e4