OBS: let osmocom-keyring not depend on other pgks
The osmocom-latest, osmocom-nightly etc. meta-packages depend on the
osmocom-keyring package. Fix that the osmocom-keyring package was also
depending on these meta packages.

This fixes that when uninstalling all osmocom:latest packages by
removing the osmocom-latest meta package, the osmocom-keyring package
would also get removed. This is undesirable for switching between latest
and nightly, as the osmocom-keyring package must remain installed to be
able to install new packages from the repository (as it has the repo
key).

Change-Id: If8fbbc40ee45e603de281003db2f61dcb7603e7b