OBS: configure Osmocom pyhss fork
Add the Osmocom fork of PyHSS to scripts/obs/lib/config.py, so the
Osmocom_OBS_* jobs build source packages for it and push them to our OBS
where binary packages get built. We need to use a fork for now, as the
official repository can't run outside of a git tree yet and doesn't have
debian packaging. These changes are being upstreamed, the goal is that
we don't need to use our fork in the long run.

Related: SYS#6819
Change-Id: I8fc7ef9dc242248770711da2893492fae8a84522